Brand new version of test, now in Marionette
