+++ Gemfile
@@ -1,35 +1,39 @@
 source 'https://rubygems.org'
 
+ruby "2.2.2"
 
-# Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
 gem 'rails', '4.1.7'
-# Use sqlite3 as the database for Active Record
 gem 'sqlite3'
-# Use SCSS for stylesheets
 gem 'sass-rails', '~> 4.0.3'
-# Use Uglifier as compressor for JavaScript assets
 gem 'uglifier', '>= 1.3.0'
-# Use CoffeeScript for .js.coffee assets and views
 gem 'coffee-rails', '~> 4.0.0'
-# See https://github.com/sstephenson/execjs#readme for more supported runtimes
-# gem 'therubyracer',  platforms: :ruby
-
-# Use jquery as the JavaScript library
-gem 'jquery-rails'
-# Turbolinks makes following links in your web application faster. Read more: https://github.com/rails/turbolinks
-gem 'turbolinks'
-# Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
 gem 'jbuilder', '~> 2.0'
-# bundle exec rake doc:rails generates the API under doc/api.
-gem 'sdoc', '~> 0.4.0', group: :doc
-
-# Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
 gem 'spring', group: :development
-
-gem 'bootstrap-sass', '~> 3.3.4'
-
 gem 'haml-rails'
 
+# Client-side libraries
+gem 'jquery-rails'
+gem 'bootstrap-sass', '~> 3.3.4'
+gem 'backbone-rails'
+gem 'marionette-rails'
+gem 'modernizr-rails'
+gem 'ejs'
+
 group :development, :test do
+  # Testing
+  gem 'konacha'
+  gem 'poltergeist'
   gem 'rspec-rails'
+  gem 'sinon-chai-rails'
+  gem 'sinon-rails'
+  gem 'chai-jquery-rails'
+
+  # Better console
+  gem 'pry'
+  gem 'pry-byebug'
+  gem 'awesome_print'
+  gem 'hirb'
+  gem 'pry-rails'
+  gem 'coolline'
+  gem 'coderay'
 end
