+++ app/assets/javascripts/application.js.coffee
@@ -1,5 +1,8 @@
 #= require jquery
 #= require jquery_ujs
 #= require bootstrap
-#= require turbolinks
+#= require underscore
+#= require modernizr
+#= require backbone
+#= require backbone.marionette
 #= require_tree .
