+++ app/assets/javascripts/modules/questions/templates/edit_question.jst.ejs
@@ -1,0 +1,16 @@
+<form>
+  <div class="alert alert-danger hide">
+  </div>
+
+  <div class="form-group">
+    <label for="question">Question</label>
+    <input type="text" name="question" id="question" class="form-control" value="<%= question %>" />
+  </div>
+
+  <div class="form-group">
+    <label for="answer">Answer</label>
+    <input type="text" name="answer" id="answer" class="form-control" value="<%= answer %>" />
+  </div>
+
+  <input type="submit" class="btn btn-primary" value="Save" />
+</form>
