+++ app/assets/javascripts/modules/questions/templates/question.jst.ejs
@@ -1,0 +1,13 @@
+<h1>Question time...</h1>
+
+<div class="well">
+  <p><%= question %></p>
+  <div class="alert alert-success hide"></div>
+
+  <form>
+    <div class="form-group">
+      <input type="text" name="answer" class="form-control" />
+    </div>
+    <input type="submit" value="Submit answer" class="btn btn-default" />
+  </form>
+</div>
