+++ app/assets/javascripts/modules/questions/templates/questions.jst.ejs
@@ -1,0 +1,8 @@
+<h1>Choose a question!</h1>
+
+<div class="well">
+  <ul class="questions">
+  </ul>
+</div>
+
+<a href="#" class="new">Add a new question</a>
