+++ app/assets/javascripts/modules/questions/templates/questions_item.jst.ejs
@@ -1,0 +1,4 @@
+<a href="#" class="link"><%= question %></a>
+<span class="small">
+  (<a href="#" class="edit">edit</a>)
+</span>
