+++ app/assets/javascripts/renderer.js.coffee
@@ -1,0 +1,11 @@
+_.extend Marionette.Renderer,
+
+  lookups: ["modules/"]
+
+  render: (template, data) ->
+    @getTemplate(template)(data)
+
+  getTemplate: (template) ->
+    for lookup in @lookups
+      path = lookup + template
+      return JST[path] if JST[path]
