--- app/helpers/questions_helper.rb
@@ -1,2 +1,0 @@
-module QuestionsHelper
-end
