+++ app/views/api/questions/_show.jbuilder
@@ -1,0 +1,1 @@
+json.extract! question, :id, :question, :answer
