+++ app/views/api/questions/index.jbuilder
@@ -1,0 +1,1 @@
+json.array! @questions, partial: 'show', as: :question
