+++ app/views/api/questions/show.jbuilder
@@ -1,0 +1,1 @@
+json.partial! 'show', question: @question
