+++ app/views/application/index.html.haml
@@ -1,0 +1,1 @@
+.container#content
