+++ app/views/layouts/application.html.haml
@@ -2,8 +2,8 @@
 %html
   %head
     %title Quiz Master
-    = stylesheet_link_tag    'application', media: 'all', 'data-turbolinks-track' => true
-    = javascript_include_tag 'application', 'data-turbolinks-track' => true
+    = stylesheet_link_tag    'application', media: 'all'
+    = javascript_include_tag 'application'
     = csrf_meta_tags
 
   %body
