--- app/views/questions/edit.html.haml
@@ -1,2 +1,0 @@
-.container
-  = render 'form'
