--- app/views/questions/index.html.haml
@@ -1,12 +1,0 @@
-.container
-  %h1 Choose a question!
-
-  .well
-    %ul
-      - @questions.each do |question|
-        %li
-          = link_to question.question, question
-          %span.small
-            (#{link_to "edit", edit_question_path(question)})
-
-  = link_to "Add a new question", new_question_path
