--- app/views/questions/new.html.haml
@@ -1,2 +1,0 @@
-.container
-  = render 'form'
