--- app/views/questions/show.html.haml
@@ -1,11 +1,0 @@
-.container
-  %h1 Question time...
-
-  .well
-    %p= @question.question
-
-    = form_for :answer, url: answer_question_path(@question) do |f|
-      .form-group
-        = f.text_field :answer, class: 'form-control'
-
-      = f.submit "Submit answer", class: 'btn btn-default'
