+++ config/initializers/konacha.rb
@@ -1,0 +1,4 @@
+Konacha.configure do |config|
+  require 'capybara/poltergeist'
+  config.driver = :poltergeist
+end if defined?(Konacha)
