+++ spec/controllers/api/questions_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe Api::QuestionsController, :type => :controller do
+  it 'is missing specs'
+end
