--- spec/controllers/questions_controller_spec.rb
@@ -1,5 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe QuestionsController, :type => :controller do
-  it 'is missing specs'
-end
